# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Сервер мессенджера на Go. Поддерживает: личные и групповые чаты, голосовые и видеозвонки, stories, голосовые комнаты, каналы, стикеры, опросы, ботов, геолокацию, кастомные эмодзи, эффекты сообщений, избранное и многое другое.",
        "title": "Chat Messenger API",
        "termsOfService": "http://localhost:8080/terms",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8080/support",
            "email": "support@chatserver.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/contacts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/contactdomain.ContactResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/contacts/photo": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "parameters": [
                    {
                        "description": "Phone and photo URL",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contactdomain.UpdateContactPhotoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/contacts/registered": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/userdomain.UserResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/contacts/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone number or partial to search",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/contactdomain.ContactResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing query",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/contacts/sync": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "parameters": [
                    {
                        "description": "Contact list to sync",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contactdomain.SyncContactsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/emojis": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает все кастомные эмодзи, загруженные всеми пользователями. Подходит для общей галереи эмодзи в клиенте.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Кастомные эмодзи"
                ],
                "responses": {
                    "200": {
                        "description": "Массив всех кастомных эмодзи на сервере",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/emojidomain.CustomEmojiResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Позволяет загрузить собственное изображение эмодзи и задать ему короткое имя (shortcode). После загрузки эмодзи можно использовать в чатах. Поддерживаются форматы PNG, JPG, GIF (включая анимированные). Файл сохраняется в директории uploads/emojis/.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Кастомные эмодзи"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Короткое имя для эмодзи (например 'party', 'thumbsup', 'cat'). Будет использоваться в клиенте для вызова эмодзи.",
                        "name": "shortcode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Файл изображения эмодзи (PNG, JPG, GIF). Рекомендуемый размер: 128x128 пикселей.",
                        "name": "emoji",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Эмодзи успешно создан, возвращает ID, shortcode и URL файла",
                        "schema": {
                            "$ref": "#/definitions/emojidomain.CustomEmojiResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка: файл не загружен, shortcode пустой или формат не поддерживается",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/emojis/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает только те эмодзи, которые были загружены текущим пользователем. Для просмотра всех публичных эмодзи используйте GET /emojis.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Кастомные эмодзи"
                ],
                "responses": {
                    "200": {
                        "description": "Массив кастомных эмодзи пользователя с ID, shortcode и URL",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/emojidomain.CustomEmojiResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/emojis/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет эмодзи из базы данных и файл изображения с диска. Только владелец эмодзи может его удалить. После удаления эмодзи перестанет отображаться в галерее.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Кастомные эмодзи"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID эмодзи для удаления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Эмодзи успешно удалён",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка: эмодзи не найден или нет прав на удаление",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "contactdomain.ContactInput": {
            "type": "object",
            "required": [
                "name",
                "phone"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "contactdomain.ContactResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "photoUrl": {
                    "type": "string"
                },
                "userIdRef": {
                    "type": "string"
                }
            }
        },
        "contactdomain.SyncContactsRequest": {
            "type": "object",
            "required": [
                "contacts"
            ],
            "properties": {
                "contacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contactdomain.ContactInput"
                    }
                }
            }
        },
        "contactdomain.UpdateContactPhotoRequest": {
            "type": "object",
            "required": [
                "phone",
                "photoUrl"
            ],
            "properties": {
                "phone": {
                    "type": "string"
                },
                "photoUrl": {
                    "type": "string"
                }
            }
        },
...
        "emojidomain.CustomEmojiResponse": {
            "type": "object",
            "properties": {
                "animated": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "fileUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "shortcode": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "ERROR_CODE"
                },
                "error": {
                    "type": "string",
                    "example": "error description"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "response.MessageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "userdomain.UserResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isAdmin": {
                    "type": "boolean"
                },
                "lastSeen": {
                    "type": "string"
                },
                "online": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите JWT токен в формате: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}